services/catalog/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
httpx==0.27.2
tenacity==8.5.0
redis==5.1.0
celery[redis]==5.4.0
lxml==5.3.0
cryptography==43.0.3
